core: refactored params for better type checking
This is a refactoring which enables better type checking for parameters.
The main change is that the type now needs to be passed to set_param
which means the user of the API needs to know the type of a param that
should be changed beforehand.

This means that the actual type checking can happen inside the
MAVLinkParameters class for both params that are set and params that are
gotten.